fix: change name to share button
