--- apps/journeys-admin/src/components/Editor/EditToolbar/Menu/JourneyLink/index.ts
@@ -1,1 +1,0 @@
-export { JourneyLink } from './JourneyLink'
