+++ apps/journeys-admin/src/components/Editor/EditToolbar/ShareButton/EmbedJourneyDialog/EmbedJourneyDialog.stories.tsx
@@ -4,8 +4,8 @@ import { ReactElement, useState } from 'react'
 
 import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
 
-import { journeysAdminConfig } from '../../../../../../libs/storybook'
-import { publishedJourney } from '../../../../data'
+import { journeysAdminConfig } from '../../../../../libs/storybook'
+import { publishedJourney } from '../../../data'
 
 import { EmbedJourneyDialog } from './EmbedJourneyDialog'
