+++ apps/journeys-admin/src/components/Editor/EditToolbar/ShareButton/index.ts
@@ -1,0 +1,1 @@
+export { ShareButton } from './ShareButton'
