--- apps/journeys-admin/src/components/Editor/Properties/AccessControl/index.ts
@@ -1,1 +1,0 @@
-export { AccessControl } from './AccessControl'
