--- apps/journeys-admin/src/components/Editor/Properties/JourneyDetails/Language/index.ts
@@ -1,1 +1,0 @@
-export { Language } from './Language'
