--- apps/journeys-admin/src/components/Editor/Properties/JourneyDetails/index.ts
@@ -1,1 +1,0 @@
-export { JourneyDetails } from './JourneyDetails'
