--- apps/journeys-admin/src/components/Editor/Properties/index.ts
@@ -1,1 +1,0 @@
-export { Properties } from './Properties'
